Endret fra Double til BigDecimal i InntektUtil

+++ src/main/kotlin/no/nav/bidrag/beregn/felles/inntekt/inntekt.kt
@@ -2,9 +2,10 @@ package no.nav.bidrag.beregn.felles.inntekt
 
 import no.nav.bidrag.beregn.felles.bo.Periode
 import no.nav.bidrag.beregn.felles.enums.InntektType
+import java.math.BigDecimal
 
 data class InntektGrunnlag(
     val inntektDatoFraTil: Periode,
     val inntektType: InntektType,
-    val inntektBelop: Double
+    val inntektBelop: BigDecimal
 )
